Issue/693: refactor: Moved creation strategy code to `business/document/creation` directory
