+++ presentation/sheet/actor/part/abilities/momentum/actor-momentum-viewmodel.mjs
@@ -1,7 +1,7 @@
 import { ITEM_TYPES } from "../../../../../../business/document/item/item-types.mjs";
 import { StringUtil } from "../../../../../../business/util/string-utility.mjs";
 import { ValidationUtil } from "../../../../../../business/util/validation-utility.mjs";
-import SpecificDocumentCreationStrategy from "../../../../../component/button-add/specific-document-creation-strategy.mjs";
+import SpecificDocumentCreationStrategy from "../../../../../../business/document/creation/specific-document-creation-strategy.mjs";
 import CompositeSortableListViewModel from "../../../../../component/composite-sortable-list/composite-sortable-list-viewmodel.mjs";
 import { SortingOption } from "../../../../../component/sort-controls/sort-controls-viewmodel.mjs";
 import DocumentListItemOrderDataSource from "../../../../../component/sortable-list/document-list-item-order-datasource.mjs";
