+++ presentation/sheet/item/expertise/expertise-table-viewmodel.mjs
@@ -6,7 +6,7 @@ import { StringUtil } from "../../../../business/util/string-utility.mjs"
 import { UuidUtil } from "../../../../business/util/uuid-utility.mjs"
 import { ValidationUtil } from "../../../../business/util/validation-utility.mjs"
 import ButtonAddViewModel from "../../../component/button-add/button-add-viewmodel.mjs"
-import ExpertiseCreationStrategy from "../../../component/button-add/expertise-creation-strategy.mjs"
+import ExpertiseCreationStrategy from "../../../../business/document/creation/expertise-creation-strategy.mjs"
 import ButtonToggleVisibilityViewModel from "../../../component/button-toggle-visibility/button-toggle-visibility-viewmodel.mjs"
 import ButtonViewModel from "../../../component/button/button-viewmodel.mjs"
 import ListFooterViewModel from "../../../component/sortable-list/list-footer-viewmodel.mjs"
